+++ cogs/notes.py
@@ -50,7 +50,7 @@ class Notes(commands.Cog):
         with open(f"user_notes/{ctx.message.author.display_name}.txt", "w") as n:
             n.write(clear_message)
         await Harbinger.send_dm(
-            ctx=ctx, member=ctx.message.author, content=f"Cleared your notes!"
+            ctx=ctx, member=ctx.message.author, content="Cleared your notes!"
         )
         Harbinger.timestamp(ctx.message.author, cmd, cmd_msg)